mvp/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
requests>=2.31
faster-whisper>=1.0
python-dotenv>=1.0
